# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Esta é a API de gerenciamento de livros da Biblioteca.",
        "title": "API da Biblioteca",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
...
        "/api/books": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Recebe um objeto JSON BookRequest e salva o livro no banco de dados.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "books"
                ],
                "summary": "Cria um novo livro",
                "parameters": [
                    {
                        "description": "Dados do Novo Livro a ser criado",
                        "name": "book",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/books.BookRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Livro criado com sucesso",
                        "schema": {
                            "$ref": "#/definitions/books.BookRequest"
                        }
                    },
                    "400": {
                        "description": "Requisição Inválida (JSON malformado ou campo obrigatório ausente)",
                        "schema": {
                            "$ref": "#/definitions/middleware.APIError"
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "schema": {
                            "$ref": "#/definitions/middleware.APIError"
                        }
                    }
                }
            }
        },
        "/api/books/relation": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Cria uma relação entre um livro existente e uma categoria existente.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "books"
                ],
                "summary": "Associa um livro a uma categoria",
                "parameters": [
                    {
                        "description": "IDs do Livro e da Categoria a serem relacionados",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/books.BookCategoryRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK, Relação criada com sucesso"
                    },
                    "400": {
                        "description": "Requisição Inválida (JSON malformado, campo ausente ou tipo errado)",
                        "schema": {
                            "$ref": "#/definitions/middleware.APIError"
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor (Erro na criação da relação)",
                        "schema": {
                            "$ref": "#/definitions/middleware.APIError"
                        }
                    }
                }
            }
        },
        "/api/books/{id}": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Recebe um objeto JSON BookRequest e atualiza o livro no banco de dados.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "books"
                ],
                "summary": "Atualiza um livro",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Recebe o id do livro",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Dados do Novo Livro a ser atualizado",
                        "name": "book",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/books.BookRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Livro atualizado com sucesso",
                        "schema": {
                            "$ref": "#/definitions/books.BookRequest"
                        }
                    },
                    "400": {
                        "description": "Requisição Inválida (JSON malformado ou campo obrigatório ausente)",
                        "schema": {
                            "$ref": "#/definitions/middleware.APIError"
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "schema": {
                            "$ref": "#/definitions/middleware.APIError"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Exclui um livro específico do banco de dados.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "books"
                ],
                "summary": "Exclui um livro pelo ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID do Livro a ser excluído",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Nenhum Conteúdo"
                    },
                    "400": {
                        "description": "Requisição Inválida (ID com formato incorreto)",
                        "schema": {
                            "$ref": "#/definitions/middleware.APIError"
                        }
                    },
                    "404": {
                        "description": "Livro não encontrado",
                        "schema": {
                            "$ref": "#/definitions/middleware.APIError"
                        }
                    }
                }
            }
        },
...
        "/public/api/books": {
            "get": {
                "description": "Retorna uma lista de livros filtrando por título, autor e categoria.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "books"
                ],
                "summary": "Listar livros",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Página",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Filtrar por título",
                        "name": "title",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filtrar por autor",
                        "name": "author",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filtrar por categoria",
                        "name": "category",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/books.BookResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Parâmetros inválidos",
                        "schema": {
                            "$ref": "#/definitions/middleware.APIError"
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "schema": {
                            "$ref": "#/definitions/middleware.APIError"
                        }
                    }
                }
            }
        },
        "/public/api/books/{id}": {
            "get": {
                "description": "Retorna um livro específico.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "books"
                ],
                "summary": "Obter livro",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID do livro",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/books.BookResponse"
                        }
                    },
                    "400": {
                        "description": "ID inválido",
                        "schema": {
                            "$ref": "#/definitions/middleware.APIError"
                        }
                    },
                    "404": {
                        "description": "Livro não encontrado",
                        "schema": {
                            "$ref": "#/definitions/middleware.APIError"
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "schema": {
                            "$ref": "#/definitions/middleware.APIError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "authors.Authors": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer",
                    "format": "int64"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "books.BookCategoryRequest": {
            "description": "Dados necessários pra fazer o relacionamento",
            "type": "object",
            "required": [
                "book_id",
                "category_id"
            ],
            "properties": {
                "book_id": {
                    "type": "integer",
                    "example": 1
                },
                "category_id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "books.BookRequest": {
            "description": "Dados necessários para criar um livro",
            "type": "object",
            "required": [
                "author_id",
                "content",
                "description",
                "title"
            ],
            "properties": {
                "author_id": {
                    "type": "integer",
                    "example": 1
                },
                "content": {
                    "type": "string",
                    "example": "A menina é o porquinho"
                },
                "description": {
                    "type": "string",
                    "example": "Esse livro é sobre conteudo infantil"
                },
                "title": {
                    "type": "string",
                    "example": "A menina e o porquinho"
                }
            }
        },
        "books.BookResponse": {
            "type": "object",
            "properties": {
                "author": {
                    "$ref": "#/definitions/authors.Authors"
                },
                "author_id": {
                    "type": "integer"
                },
                "categories": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/categories.Category"
                    }
                },
                "content": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "categories.Category": {
            "type": "object",
            "properties": {
                "createdAT": {
                    "type": "string"
                },
                "id": {
                    "type": "integer",
                    "format": "int64"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "middleware.APIError": {
            "description": "Erro padronizado de resposta.",
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}